% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_eagles.R
\name{hello_eagles}
\alias{hello_eagles}
\title{Say Hello to the EAGLES
This function says Hello to the EAGLES depending on data}
\usage{
hello_eagles(eagles, daytime = "auto")
}
\description{
@param eagles character, the names of the EAGLES to be greeted
 @param daytime character, optional, either "morning" or "evening"
 @return Nothing, it will say hello
 @importFrom lubridate hour
 @export
}
