% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2.R
\name{hello_eagles}
\alias{hello_eagles}
\title{say hello to the EAGLES}
\usage{
hello_eagles(eagles, daytime = "auto")
}
\arguments{
\item{eagles}{character, the names of the EAGLEs to be greeted}

\item{daytime}{character, optional, either "auto" ....}
}
\value{
Nothing, it will just say "hello"
}
\description{
this function say hello to EAGLES depending on daytime
}
