% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating packages.R
\name{hello_eagles}
\alias{hello_eagles}
\title{Say hello to the eagles}
\usage{
hello_eagles(eagles, daytime = "auto")
}
\arguments{
\item{eagles}{character. The names of the Eagles to be greeted}

\item{daytime}{character. optional, either "auto" to calculate daytime or "morning" or "afternoon" or "evening"}
}
\value{
Nothing, it will just say hello
}
\description{
This function says hello to Eagles depending on daytime
}
