% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_eagles.R
\name{hello_eagles}
\alias{hello_eagles}
\title{Say Hello to the EAGLES}
\usage{
hello_eagles(eagles, daytime = "auto")
}
\arguments{
\item{eagles}{character, the names of the EAGLES to be greeted}

\item{daytime}{character, optional, either "morning" or "evening"}
}
\value{
Nothing, just say hello
}
\description{
This function says Hello to the EAGLES depending on data
}
